backend/requirements/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
aniso8601==7.0.0          # via graphene
appdirs==1.4.4            # via black
asgiref==3.3.1            # via django
black==20.8b1             # via -r dev-requirements.in
brotli==1.0.9             # via whitenoise
click==7.1.2              # via black
dj-database-url==0.5.0    # via django-heroku
django-graphql-jwt==0.3.1  # via -r requirements.in
django-heroku==0.3.1      # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-graphql-jwt, django-heroku, graphene-django
graphene-django==2.13.0   # via -r requirements.in, django-graphql-jwt
graphene==2.1.8           # via -r requirements.in, graphene-django
graphql-core==2.3.2       # via django-graphql-jwt, graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via graphene
gunicorn==20.0.4          # via -r requirements.in
mypy-extensions==0.4.3    # via black
pathspec==0.8.1           # via black
promise==2.3              # via graphene-django, graphql-core, graphql-relay
psycopg2==2.8.6           # via django-heroku
pyjwt==1.7.1              # via django-graphql-jwt
pytz==2020.4              # via django
regex==2020.11.13         # via black
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.15.0               # via graphene, graphene-django, graphql-core, graphql-relay, promise, singledispatch
sqlparse==0.4.1           # via django
toml==0.10.2              # via black
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
unidecode==1.1.2          # via graphene-django
whitenoise[brotli]==5.2.0  # via -r requirements.in, django-heroku

# The following packages are considered to be unsafe in a requirements file:
# setuptools
